refactor: 영수증 정보 수정 navbar 수정
