+++ src/components/common/Navbar/Navbar.tsx
@@ -4,8 +4,12 @@ import classNames from "classnames";
 
 import styles from "@/components/common/Navbar/Navbar.module.scss";
 
-const Navbar = ({ children }: PropsWithChildren) => {
-  return <div className={styles.Navbar}>{children}</div>;
+interface NavbarProps extends PropsWithChildren {
+  className?: string;
+}
+
+const Navbar = ({ children, className }: NavbarProps) => {
+  return <div className={(classNames(styles.Navbar), className)}>{children}</div>;
 };
 
 Navbar.LeftButton = ({
