+++ src/pages/ReceiptEditPage/ReceiptEditPage.module.scss
@@ -31,6 +31,15 @@
   }
 }
 
+.Navbar {
+  padding-left: 0;
+  padding-right: 0;
+}
+
+.Container {
+  z-index: 1;
+}
+
 .Top {
   z-index: 1;
